% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFfe.R
\name{BFfe}
\alias{BFfe}
\title{Shiny Application for Setting Up and Running Bayes Factors for Each Participant}
\usage{
BFfe()
}
\value{
Launches a Shiny application in the user's default web browser.
}
\description{
This function launches a 'shiny' application that allows users to set up and
run a Bayes Factors analysis for each participant using the \code{BF_for_everyone}
function.
}
\details{
The app allows users to upload a CSV file, specify the formula for the
linear model, define hypotheses, and select the participant variable. The app
then runs the analysis and displays the results, including a summary of Bayes
Factors, geometric product of Bayes Factors, and individual participant results.
}
\examples{
# To run the Shiny app
if(interactive()){
  BFfe()
}

}
\references{
Klaassen, F. (2020). Combining Evidence Over Multiple Individual Analyses. In
R. van de Schoot & M. Miočević (Eds.), Small Sample Size Solutions: A Guide for
Applied Researchers and Practitioners (1st ed., pp. 13). Routledge.
\url{doi:10.4324/9780429273872-11}
}
