% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepliCrisis.R
\name{RepliCrisis}
\alias{RepliCrisis}
\title{RepliCrisis Shiny App}
\usage{
RepliCrisis()
}
\value{
An interactive Shiny app.
}
\description{
The RepliCrisis Shiny app is an interactive game based on the process for
evaluating replication studies presented in Hoijtink et. al. (2009). Users can
conduct original studies, process the results, and attempt to replicate the
studies while adjusting parameters to see how changes affect the outcome.
}
\examples{
if(interactive()){
RepliCrisis()
}

}
\references{
Hoijtink, H., Mulder, J., van Lissa, C., & Gu, X. (2019). A tutorial on
testing hypotheses using the Bayes factor. Psychological methods, 24(5),
539–556. \url{doi:10.1037/met0000201}
}
\keyword{shinyapp}
